move SessionId to separate module
